# laravel/storage/api-docs/swagger.yaml
openapi: 3.0.3
info:
  title: API
  version: 1.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /api/v1/product/types:
    get:
      tags:
        - PRODUCT TYPES
      summary: Get all product types
      description: Retrieves a list of all product types available in the system.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Product types fetched successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Product types fetched successfully.
                  errors:
                    type: array
                    items: {}
                    example: []
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        name:
                          type: string
                          example: "Roupas"
                        slug:
                          type: string
                          example: "roupas"
                        description:
                          type: string
                          example: "Todas as peças de vestuário"
                        order:
                          type: integer
                          example: 1
                        icon:
                          type: string
                          nullable: true
                          example: null
                        image_url:
                          type: string
                          nullable: true
                          example: null
                        active:
                          type: integer
                          example: 1
                        created_at:
                          type: string
                          example: "2026-02-11 10:25:12"
                        updated_at:
                          type: string
                          example: "2026-02-11 10:25:12"
                    example:
                      - id: 1
                        name: "Roupas"
                        slug: "roupas"
                        description: "Todas as peças de vestuário"
                        order: 1
                        icon: null
                        image_url: null
                        active: 1
                        created_at: "2026-02-11 10:25:12"
                        updated_at: "2026-02-11 10:25:12"
                      - id: 10
                        name: "Eletrônicos"
                        slug: "eletronicos"
                        description: "Produtos eletrônicos e tecnologia"
                        order: 2
                        icon: null
                        image_url: null
                        active: 1
                        created_at: "2026-02-11 10:25:12"
                        updated_at: "2026-02-11 10:25:12"
                      - id: 20
                        name: "Móveis"
                        slug: "moveis"
                        description: "Móveis para casa e escritório"
                        order: 3
                        icon: null
                        image_url: null
                        active: 1
                        created_at: "2026-02-11 10:25:12"
                        updated_at: "2026-02-11 10:25:12"
                      - id: 30
                        name: "Livros"
                        slug: "livros"
                        description: "Livros e publicações"
                        order: 4
                        icon: null
                        image_url: null
                        active: 1
                        created_at: "2026-02-11 10:25:12"
                        updated_at: "2026-02-11 10:25:12"
                  _links:
                    type: object
                    properties:
                      self:
                        type: object
                        properties:
                          href:
                            type: string
                            example: http://localhost:8020/product/types
  /api/v1/product/type/{id}:
    get:
      tags:
        - PRODUCT TYPES
      summary: Get product type by ID
      description: Retrieves a single product type by its ID.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the product type to retrieve
          schema:
            type: integer
            example: 2
      responses:
        '200':
          description: Product type fetched successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Product type fetched successfully.
                  errors:
                    type: array
                    items: {}
                    example: []
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 2
                      name:
                        type: string
                        example: "Camisas"
                      slug:
                        type: string
                        example: "camisas"
                      description:
                        type: string
                        example: "Camisas masculinas e femininas"
                      parent_id:
                        type: integer
                        nullable: true
                        example: 1
                      variant_type:
                        type: string
                        example: "clothing"
                      order:
                        type: integer
                        example: 1
                      icon:
                        type: string
                        nullable: true
                        example: null
                      image_url:
                        type: string
                        nullable: true
                        example: null
                      active:
                        type: integer
                        example: 1
                      created_at:
                        type: string
                        example: "2026-02-11 10:25:12"
                      updated_at:
                        type: string
                        example: "2026-02-11 10:25:12"
                  _links:
                    type: object
                    properties:
                      self:
                        type: object
                        properties:
                          href:
                            type: string
                            example: http://localhost:8020/v1/product/type/2
                      GET_ALL:
                        type: object
                        properties:
                          href:
                            type: string
                            example: http://localhost:8020/v1/product/types
        '422':
          description: Invalid parameters.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Error Processing Request
                  errors:
                    type: array
                    items:
                      type: string
                    example:
                      - Invalid id parameter
                  data:
                    type: array
                    example: []
        '401':
          description: Unauthorized. Bearer token is missing, expired, or invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Token not found
                  errors:
                    type: array
                    items: {}
                    example: []
                  data:
                    type: array
                    example: []
        '404':
          description: Product type not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Product type not found.
                  errors:
                    type: array
                    items: {}
                    example: []
                  data:
                    type: array
                    example: []
    delete:
      tags:
        - PRODUCT TYPES
      summary: Delete product type
      description: Soft deletes a product type by its ID.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the product type to delete
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Product type deleted successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Product type deleted successfully.
                  errors:
                    type: array
                    items: {}
                    example: []
                  data:
                    type: boolean
                    example: true
                  _links:
                    type: object
                    properties:
                      self:
                        type: object
                        properties:
                          href:
                            type: string
                            example: http://localhost:8020/v1/product/type/1/status
                      GET_TYPE:
                        type: object
                        properties:
                          href:
                            type: string
                            example: http://localhost:8020/v1/product/type/1
                      GET_ALL:
                        type: object
                        properties:
                          href:
                            type: string
                            example: http://localhost:8020/v1/product/types
                      GET_CHILD:
                        type: object
                        properties:
                          href:
                            type: string
                            example: http://localhost:8020/v1/product/type/1/child
        '401':
          description: Unauthorized. Bearer token is missing, expired, or invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Token not found
                  errors:
                    type: array
                    items: {}
                    example: []
                  data:
                    type: array
                    items: {}
                    example: []
        '404':
          description: Product type not found or already deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Parent id not found
                  errors:
                    type: array
                    items:
                      type: string
                    example:
                      - There is no parent with id 1.
                  data:
                    type: array
                    items: {}
                    example: []
        '500':
          description: Persistence error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Error processing request.
                  errors:
                    type: array
                    items: {}
                    example: []
                  data:
                    type: array
                    items: {}
                    example: []
        '422':
          description: Error validating fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Error Processing Request"
                  errors:
                    type: array
                    items:
                      type: string
                    example:
                      - "The email field is required."
                  data:
                    type: array
                    items: {}
                    example: []
...
  /api/v1/product/type/{id}/status:
    patch:
      tags:
        - PRODUCT TYPES
      summary: Change product type activation status
      description: Activates or deactivates a product type by its ID.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the product type
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [TRUE, FALSE]
                  example: "TRUE"
      responses:
        '200':
          description: Product type updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Product type updated successfully.
                  errors:
                    type: array
                    items: {}
                    example: []
                  data:
                    type: boolean
                    example: true
                  _links:
                    type: object
                    properties:
                      self:
                        type: object
                        properties:
                          href:
                            type: string
                            example: http://localhost:8020/v1/product/type/1/status
                      GET_TYPE:
                        type: object
                        properties:
                          href:
                            type: string
                            example: http://localhost:8020/v1/product/type/1
                      GET_ALL:
                        type: object
                        properties:
                          href:
                            type: string
                            example: http://localhost:8020/v1/product/types
                      GET_CHILD:
                        type: object
                        properties:
                          href:
                            type: string
                            example: http://localhost:8020/v1/product/type/1/child
        '422':
          description: Invalid parameters.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Invalid parameters.
                  errors:
                    type: array
                    items:
                      type: string
                    example:
                      - ID is not valid.
                  data:
                    type: array
                    items: {}
                    example: []
        '401':
          description: Unauthorized. Bearer token is missing, expired, or invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Token not found
                  errors:
                    type: array
                    items: {}
                    example: []
                  data:
                    type: array
                    items: {}
                    example: []
        '404':
          description: Product type not found or deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Parent id not found
                  errors:
                    type: array
                    items:
                      type: string
                    example:
                      - There is no parent with id 1.
                  data:
                    type: array
                    items: {}
                    example: []